requirements.txt
django==4.2.4
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
Pillow==10.0.0
environs[django]==9.5.0
django-phonenumber-field==7.1.0

